The avatar will be moving on the path between the two points on the map
Given 2 points (latitude and longitude), the objective is to guide an avatar along the path formed by
these points,ensuring the avatar follows the entire path without interruptions. The avatar's journey
begins at the initial point and progresses through subsequent points to reach the final endpoint, all to be
displayed on a map. The journey must take a minimum of 30 seconds.
Sample Case:
Point 1: lat: 28.7565, lng: 77.756578
Point 2: lat: 28.9565, lng: 77.706578
Outcome:
The avatar will be moving on the path between the two points on the map